+++ src/lib.rs
@@ -118,7 +118,11 @@
 
 pub mod future;
 pub mod stream;
-pub mod time;
+
+#[cfg(any(feature = "async-io", feature = "docs"))]
+pub mod async_io;
+#[cfg(feature = "tokio")]
+pub mod tokio;
 
 mod deadline;
 mod stop_source;
